requirement.txt
Flask==3.0.2
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
gunicorn==21.2.0
